Remove VFS Swift layout v1 and v2
They have been deprecated a few years ago, and the V3 layout is really
better.
